refactor(plugins): extract route helpers into route_utils.py in googledrive
Moves get_drive_credentials and _sync_folder_files out of dependencies.py
into a new route_utils.py so that dependencies.py only contains true
FastAPI Depends-injectable functions.
