+++ app/core_plugins/googledrive/routes/folders.py
@@ -8,12 +8,8 @@ from sqlmodel import Session, select
 
 from app.core.db import get_session
 from app.core_plugins.googledrive.client import GoogleDriveClient
-from app.core_plugins.googledrive.routes.dependencies import (
-    _sync_folder_files,
-    get_drive_credentials,
-    get_valid_access_token,
-    require_user_id,
-)
+from app.core_plugins.googledrive.routes.dependencies import get_valid_access_token, require_user_id
+from app.core_plugins.googledrive.routes.route_utils import _sync_folder_files, get_drive_credentials
 from app.core_plugins.googledrive.types import (
     CreateFolderRequest,
     DriveFileResponse,
